requirements.txt
streamlit
openai
pandas
PyPDF2
python-dotenv
requests
